validating endpoints

+++ app/api/users/views/users.py
@@ -1,9 +1,8 @@
-from utils.validations import validate_users_key_pair_values
-from flask import jsonify, request
+from flask import jsonify, request, abort
 from flask.helpers import make_response
 from app.api.users import blueprint as users
 from app.api.users.models.users import User, user_schema, users_schema
-from utils.validations import validate_users_key_pair_values, error, check_for_blanks, check_for_non_strings
+from utils.validations import validate_users_key_pair_values, check_for_blanks, check_for_non_strings
 from app import db
 
 
@@ -21,6 +20,8 @@ def get():
 """
 @users.route('/user/<int:id>', methods=['GET'])
 def get_user(id):
+    if id <= 0:
+        return abort(400, "Invalid id format")
     user = User.query.get(id)
     if user:
         return user_schema.dump(user), 200
@@ -38,15 +39,15 @@ def get_user(id):
 def post():
 
     if validate_users_key_pair_values(request):
-        return error(400, "{} key missing".format(', '.join(validate_users_key_pair_values(request))))
+        return abort(400, "{} key missing".format(', '.join(validate_users_key_pair_values(request))))
 
     data = request.get_json()
 
     if check_for_blanks(data):
-        return error(400, "{} cannot be blank".format(', '.join(check_for_blanks(data))))
+        return abort(400, "{} cannot be blank".format(', '.join(check_for_blanks(data))))
 
     if check_for_non_strings(data):
-        return error(400, "{} must be a string".format(', '.join(check_for_non_strings(data))))
+        return abort(400, "{} must be a string".format(', '.join(check_for_non_strings(data))))
 
     firstName = data.get('firstName')
     lastName = data.get('lastName')
@@ -72,13 +73,26 @@ def post():
 
 
 """
-    Edit a user
+    Edit a specific user
 """
 @users.route('/edit_user/<int:id>', methods=['PUT'])
 def edit_user(id):
+    if id <= 0:
+        return abort(400, "Invalid id format")
+
     user = User.query.get(id)
 
+    if validate_users_key_pair_values(request):
+        return abort(400, "{} key missing".format(', '.join(validate_users_key_pair_values(request))))
+
     data = request.get_json()
+
+    if check_for_blanks(data):
+        return abort(400, "{} cannot be blank".format(', '.join(check_for_blanks(data))))
+
+    if check_for_non_strings(data):
+        return abort(400, "{} must be a string".format(', '.join(check_for_non_strings(data))))
+
     firstName = data.get('firstName')
     lastName = data.get('lastName')
     userName = data.get('userName')
@@ -111,6 +125,9 @@ def edit_user(id):
 """
 @users.route('/delete_user/<int:id>', methods=['DELETE'])
 def delete_user(id):
+    if id <= 0:
+        return abort(400, "Invalid id format")
+
     user = User.query.get(id)
     try:
         db.session.delete(user)
